pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "prism-challenge"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "dj-database-url (>=2.3.0,<3.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
